+++ src/constants/cardData.js
@@ -1,0 +1,67 @@
+const sampleData = [
+  {
+    name: '박유진',
+    job: '프로덕트 디자이너',
+    company: '숙명여자대학교',
+    imageUrl: '',
+    category: '비즈니스',
+  },
+  {
+    name: '한은영',
+    job: '프론트엔드 개발자',
+    company: '숙명여자대학교',
+    imageUrl: '',
+    category: '비즈니스',
+  },
+  {
+    name: '이영희',
+    job: '백엔드 개발자',
+    company: 'WELLET Corp.',
+    imageUrl: '',
+    category: '비즈니스',
+  },
+  {
+    name: '김현우',
+    job: '프론트엔드 개발자',
+    company: 'WELLET Corp.',
+    imageUrl: '',
+    category: '비즈니스',
+  },
+  {
+    name: '김디공',
+    job: '대표',
+    company: '다공 부동산',
+    imageUrl: '',
+    category: '부동산',
+  },
+  {
+    name: '홍길동',
+    job: '대표 | 공인중개사',
+    company: '동네공인중개사',
+    imageUrl: '',
+    category: '부동산',
+  },
+  {
+    name: '김인쇄',
+    job: '공인중개사',
+    company: '인쇄픽부동산',
+    imageUrl: '',
+    category: '부동산',
+  },
+  {
+    name: '박대기',
+    job: '보도본부 사회팀 / 기자',
+    company: 'KBS',
+    imageUrl: '',
+    category: '방송사',
+  },
+  {
+    name: '홍길동',
+    job: '팀장 / PD',
+    company: 'KBS VJ 특공대',
+    imageUrl: '',
+    category: '방송사',
+  },
+];
+
+export default sampleData;
