--- src/constants/cards.js
@@ -1,102 +1,0 @@
-const sampleData = [
-  {
-    name: '박유진',
-    job: '프로덕트 디자이너',
-    company: '숙명여자대학교',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '한은영',
-    job: '프론트엔드 개발자',
-    company: '숙명여자대학교',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '이영희',
-    job: '백엔드 개발자',
-    company: 'WELLET Corp.',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '순희네보리밥',
-    job: '',
-    company: '',
-    imageUrl: '',
-    category: '음식점',
-  },
-  {
-    name: '곱분이곱창',
-    job: '',
-    company: '',
-    imageUrl: '',
-    category: '음식점',
-  },
-  {
-    name: '박유진',
-    job: '프로덕트 디자이너',
-    company: '숙명여자대학교',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '한은영',
-    job: '프론트엔드 개발자',
-    company: '숙명여자대학교',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '이영희',
-    job: '백엔드 개발자',
-    company: 'WELLET Corp.',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '박유진',
-    job: '프로덕트 디자이너',
-    company: '숙명여자대학교',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '한은영',
-    job: '프론트엔드 개발자',
-    company: '숙명여자대학교',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '이영희',
-    job: '백엔드 개발자',
-    company: 'WELLET Corp.',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '박유진',
-    job: '프로덕트 디자이너',
-    company: '숙명여자대학교',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '한은영',
-    job: '프론트엔드 개발자',
-    company: '숙명여자대학교',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-  {
-    name: '이영희',
-    job: '백엔드 개발자',
-    company: 'WELLET Corp.',
-    imageUrl: '',
-    category: '비즈니스',
-  },
-];
-
-export default sampleData;
